requirements.txt
pywebview>=5.0
keyboard>=0.13.5
pystray>=0.19
Pillow>=10.0
